[python] Rename `utils.py` to `_index_util.py`
