+++ apis/python/src/tiledbsoma/_index_util.py
@@ -1,3 +1,8 @@
+"""
+This file is separate from _util.py, due to a circular-import issue with
+SOMATileDBContext which would otherwise ensue.
+"""
+
 import numpy as np
 import pandas as pd
 
@@ -5,12 +10,11 @@ from tiledbsoma import pytiledbsoma as clib
 
 from .options import SOMATileDBContext
 
-"""Builds an indexer object compatible with :meth:`pd.Index.get_indexer`."""
-
 
 def build_index(
     keys: np.typing.NDArray[np.int64], context: SOMATileDBContext
 ) -> clib.IntIndexer:
+    """Builds an indexer object compatible with :meth:`pd.Index.get_indexer`."""
     if len(np.unique(keys)) != len(keys):
         raise pd.errors.InvalidIndexError(
             "Reindexing only valid with uniquely valued Index objects"
